refactor Commits: reestruturando o código com no modelo de arquiteturas de camadas
