+++ src/main/java/com/java10x/CadastroDeNinja/Missoes/MissoesController/MissoesController.java
@@ -1,0 +1,4 @@
+package com.java10x.CadastroDeNinja.Missoes.MissoesController;
+
+public class MissoesController {
+}
