+++ src/main/java/com/java10x/CadastroDeNinja/Missoes/MissoesModel/MissoesModel.java
@@ -1,0 +1,33 @@
+package com.java10x.CadastroDeNinja.Missoes.MissoesModel;
+
+
+import com.java10x.CadastroDeNinja.Ninjas.NinjaModel.NinjaModel;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Entity
+@Table(name = "tb_missoes")
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+public class MissoesModel {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String nome;
+
+    private String grauDificuldade;
+
+    // Uma missaão para varios ninjas
+    @OneToMany(mappedBy = "missoes")
+    private List<NinjaModel> ninjas;
+
+
+
+}
