+++ src/main/java/com/java10x/CadastroDeNinja/Missoes/MissoesRepository/MissoesRepository.java
@@ -1,0 +1,7 @@
+package com.java10x.CadastroDeNinja.Missoes.MissoesRepository;
+
+import com.java10x.CadastroDeNinja.Missoes.MissoesModel.MissoesModel;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MissoesRepository extends JpaRepository<MissoesModel, Long> {
+}
