+++ src/main/java/com/java10x/CadastroDeNinja/Missoes/MissoesService/MissoesService.java
@@ -1,0 +1,4 @@
+package com.java10x.CadastroDeNinja.Missoes.MissoesService;
+
+public class MissoesService {
+}
