+++ src/main/java/com/java10x/CadastroDeNinja/Ninjas/NinjaRepository/NinjaRepository.java
@@ -1,0 +1,4 @@
+package com.java10x.CadastroDeNinja.Ninjas.NinjaRepository;
+
+public class NinjaRepository {
+}
