+++ src/main/java/com/java10x/CadastroDeNinja/Ninjas/NinjaService/NinjaService.java
@@ -1,0 +1,4 @@
+package com.java10x.CadastroDeNinja.Ninjas.NinjaService;
+
+public class NinjaService {
+}
